+++ app/src/main/java/com/sysadmindoc/guitartuner/tuning/MeasurementFreeze.kt
@@ -42,6 +42,7 @@ class MeasurementFreeze {
 
         TuningStatus.WaitingForSignal,
         TuningStatus.SignalClipping,
+        TuningStatus.HighNoise,
         TuningStatus.NoStringDetected,
         -> false
     }